requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.5.2
kubernetes==28.1.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
python-json-logger==2.0.7
aiohttp==3.9.1
python-dotenv==1.0.0
